chore: Restructured project folders
Added service classes, renamed folders and added new to match SDD, ODD
