+++ src/logic/AmbienteAgricolo/AmbienteAgricoloService.py
@@ -1,0 +1,28 @@
+from src.logic.model.Terreno import Terreno
+from src.logic.Storage.TerrenoDAO import TerrenoDAO
+
+#TODO: ERROR HANDLING DAO TERRENO
+class AmbienteAgricoloService():
+    def aggiungiTerreno(nome: str, coltura:str, posizione, preferito:bool, priorita:int)-> bool:
+        id = None
+        terreno = Terreno(id, nome, coltura, posizione, preferito, priorita)
+        TerrenoDAO.InserisciTerreno(terreno)
+        return True 
+
+    def trovaTerreno(id: str)-> Terreno:
+        Terreno = TerrenoDAO.TrovaTerreno(id)
+        return Terreno
+    
+    def modificaTerreno(id:str, nome: str, coltura:str, posizione, preferito:bool, priorita:int)-> bool:
+        terreno = Terreno(id, nome, coltura, posizione, preferito, priorita)
+        TerrenoDAO.modificaTerreno(terreno)
+        return True
+    
+    def eliminaTerreno(id:str)-> bool:
+        terreno = TerrenoDAO.TrovaTerreno(id)
+        if(terreno is None):
+            return False
+        else:
+            TerrenoDAO.RimuoviTerreno(terreno)
+            return True
+
