+++ src/logic/Storage/LicenzaDAO.py
